+++ example/main.dart
@@ -3,9 +3,10 @@ import 'dart:io';
 import 'package:freedesktop_desktop_entry/freedesktop_desktop_entry.dart';
 
 void main() async {
-  final file = File("test/desktop_entry_files/desktop-entry-1.desktop");
-  String content = await file.readAsString();
-  DesktopEntry desktopEntry = DesktopEntry.parse(content);
+  Directory.current = "test/desktop_entry_files/";
+  final file = File("desktop-entry-1.desktop");
+
+  DesktopEntry desktopEntry = await DesktopEntry.parseFile(file);
 
   LocalizedDesktopEntry localizedDesktopEntry =
       desktopEntry.localize(lang: 'fr', country: 'BE');
@@ -27,4 +28,9 @@ void main() async {
       ?.localizedValues[Locale(lang: 'en')]
       ?.getStringList();
   print(englishKeywords);
+
+  print(desktopEntry.id);
+
+  final allFiles = await parseAllInstalledDesktopFiles();
+  print(allFiles.length);
 }