+++ lib/src/localized_desktop_entry.dart
@@ -1,3 +1,4 @@
+import 'package:freedesktop_desktop_entry/src/desktop_entry.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'localized_desktop_entry.freezed.dart';
@@ -7,6 +8,7 @@ class LocalizedDesktopEntry with _$LocalizedDesktopEntry {
   const LocalizedDesktopEntry._();
 
   const factory LocalizedDesktopEntry({
+    required DesktopEntry desktopEntry,
     required Map<String, String> entries,
     @Default({}) Map<String, Map<String, String>> actions,
   }) = _LocalizedDesktopEntry;